+++ electron-vite-project/src/components/LightSettings.tsx
@@ -34,6 +34,7 @@ function LightSettings({ onClose }: SettingsProps) {
   const [inputName, setInputName] = useState('');
   const [inputDMXstart, setInputDMXstart] = useState('');
   const [inputDMXrange, setInputDMXrange] = useState('');
+  const [rangeChanged, setRangeChanged] = useState(false);
   const [inputUniverse, setInputUniverse] = useState('');
   const [inputType, setInputType] = useState('');
   const [inputNumber, setInputNumber] = useState('');
@@ -106,16 +107,11 @@ function LightSettings({ onClose }: SettingsProps) {
           }, 4000);
         }
       } else {
-        lightDeleted(); // reset selected device
+        deselectDivice();
       }
     };
     const lightDeleted = () => {
-      setSelectedDevice(undefined);
-      setIsNewDevice(false);
-      setChannelChangeArray([]);
-      setChannelArray([]);
-      setInputDMXstart('0');
-      fetchDevices();
+      deselectDivice();
     };
 
     on('light_response', lightRespone);
@@ -185,6 +181,7 @@ function LightSettings({ onClose }: SettingsProps) {
     if (!isNaN(numericValue)) {
       numericValue = Math.max(1, Math.min(513 - parseInt(inputDMXstart), numericValue));
       setInputDMXrange(Math.round(numericValue).toString());
+      setRangeChanged(true);
     } else {
       setInputDMXrange(selectedDevice?.attributes?.channel?.length?.toString() || '4'); // Reset value if input is NaN
     }
@@ -200,7 +197,7 @@ function LightSettings({ onClose }: SettingsProps) {
 
     // Set the DMX range to the number of channels of the selected device type
     const valueChannels = LampTypeChannels[value];
-    if (valueChannels) {
+    if (valueChannels && !rangeChanged) {
       setInputDMXrange(valueChannels.length.toString());
     }
   };
@@ -220,19 +217,26 @@ function LightSettings({ onClose }: SettingsProps) {
   };
 
   const handleDeselectDevice = (device: DeviceConfig) => {
+    deselectDivice();
+  };
+
+  const deselectDivice = () => {
     setSelectedDevice(undefined);
-    !isNewDevice && setUnselectedDevices([...unselectedDevices, device]);
+    //!isNewDevice && setUnselectedDevices([...unselectedDevices, device]); // instead of fetching
     setIsNewDevice(false);
     setChannelChangeArray([]);
     setChannelArray([]);
     setInputDMXstart('0');
+    setRangeChanged(false);
+    fetchDevices();
   };
 
   const handleCreateDevice = () => {
     setIsNewDevice(true);
+    const lastDevice = devices.length > 0 ? devices[devices.length - 1] : null;
 
     const newDevice: DeviceConfig = {
-      id: devices.length + 1,
+      id: Math.min(lastDevice ? lastDevice.id + 1 : 1, 691), // Last device ID + 1
       deviceValue: 0,
       name: t('ls_newDevice'),
       attributes: undefined,
@@ -246,7 +250,7 @@ function LightSettings({ onClose }: SettingsProps) {
     setInputDMXrange('4');
     setInputUniverse('U1');
     setInputType('RGBDim');
-    setInputNumber(newDevice?.id.toString() || '1');
+    setInputNumber(newDevice.id.toString() || '1');
   };
 
   // Set the initial channel array and update it